Make updating a post be about updating a request
