+++ test/helpers/react_components/student/mentoring_session_test.rb
@@ -50,7 +50,7 @@ module ReactComponents::Student
           videos: [],
           links: {
             exercise: Exercism::Routes.track_exercise_mentor_discussions_url(track, exercise),
-            create_mentor_request: Exercism::Routes.api_solution_mentor_request_path(solution.uuid),
+            create_mentor_request: Exercism::Routes.api_solution_mentor_requests_path(solution.uuid),
             learn_more_about_private_mentoring: "#",
             private_mentoring: "https://some.link/we/need/to-decide-on",
             mentoring_guide: "#"
@@ -105,7 +105,7 @@ module ReactComponents::Student
           ],
           links: {
             exercise: Exercism::Routes.track_exercise_mentor_discussions_url(track, exercise),
-            create_mentor_request: Exercism::Routes.api_solution_mentor_request_path(solution.uuid),
+            create_mentor_request: Exercism::Routes.api_solution_mentor_requests_path(solution.uuid),
             learn_more_about_private_mentoring: "#",
             private_mentoring: "https://some.link/we/need/to-decide-on",
             mentoring_guide: "#"
